MEDIUM: Fix dnf update
It was not possible to do "dnf update" to update standard system packages
because of broken dependencies. We want version restrictions on key
packages like grafana because new versions may break Kompot integration.
Kompot and key packages for which we need constraints on versions are
now hosted in a dedicated repository that will get priority.